refactor(web/chat): enhance chat functionality with new hooks and constants
- Introduced `useChatPersistence` and `useCopilotActions` hooks to manage chat thread persistence and copilot actions, respectively, improving code organization and maintainability.
- Added a new constants file for mobile breakpoint management, enhancing responsiveness.
- Updated import paths for hooks to reflect new directory structure, ensuring better modularity.
- Refactored `useIsMobile` to utilize the new mobile breakpoint constant for improved clarity.
