+++ apps/web/consts/app-consts.ts
@@ -1,0 +1,1 @@
+export const MOBILE_BREAKPOINT = 768;
